Add ssrRenderer
